add TaskCodeState, hide JOIN_TO_TEAM button after joining to team

+++ src/commands/quest.py
@@ -8,11 +8,12 @@ from telebot.states.sync import StateContext
 from database.dao import get_member, get_team_by_code, update_team, get_current_chain, get_team_by_id, join_team_via_code, get_user_ids_by_team
 from database.models import Team, Task
 from msg_locale import QuestMessages, ButtonMessages
-from buttons import render_cancel_button
+from buttons import render_cancel_button, render_main_menu
+from checks import check_admin
 
 
 
-class TaskCodeState(State):
+class TaskCodeState(StatesGroup):
     waiting_for_task_code = State()
 
 def register_quest_commands(bot: TeleBot):
@@ -42,7 +43,7 @@ def register_quest_commands(bot: TeleBot):
 
         else:
             join_team_via_code(code_word, user_id=message.from_user.id)
-            bot.reply_to(message, team.welcome_message)
+            bot.reply_to(message, team.welcome_message, reply_markup =render_main_menu(is_in_team=True))
 
         # Отправка первого задания
         current_chain = preprocess_task(message)
@@ -98,32 +99,62 @@ def register_quest_commands(bot: TeleBot):
     @bot.message_handler(func=lambda m: m.text == ButtonMessages.NEXT_TASK)
     def next_task(message: Message, state: StateContext):
         chat_id = message.chat.id
-        member = get_member(message.from_user.id)
+        user_id = message.from_user.id
+
+
+        member = get_member(user_id)
         if not member:
             bot.reply_to(message, QuestMessages.NOT_IN_TEAM)
             return
         
         
         team = get_team_by_id(member.team_id)
-
         current_chain = get_current_chain(team.id, team.current_chain_order)
+
+
         if not current_chain:
             print('not_current_chain')
             bot.reply_to(message, QuestMessages.NO_ACTIVE_TASKS)
             return
         
-        temp_data[chat_id]["team"] = team
-        temp_data[chat_id]["current_chain"] = current_chain
         state.set(TaskCodeState.waiting_for_task_code)
-        bot.reply_to(message, QuestMessages.ENTER_TASK_CODE, reply_markup=render_cancel_button())
 
+
+        temp_data[chat_id]["team_id"] = team.id
+        temp_data[chat_id]["chain_order"] = current_chain.order
+
+
+        bot.reply_to(
+            message, QuestMessages.ENTER_TASK_CODE,
+            reply_markup=render_cancel_button()
+        )
+
+
+    # Обработка ввода кодового слова
     @bot.message_handler(state=TaskCodeState.waiting_for_task_code)
-    def process_next_task(message: Message, state: StateContext, team: Team, current_chain: Task):
+    def process_next_task(message: Message, state: StateContext):
+        user_id = message.from_user.id
         chat_id = message.chat.id
-        msg = message.text
+        msg_text = message.text
         
-        state.add_data(msg)
-        check_task_code(message, temp_data[chat_id]["team"], temp_data[chat_id]["current_chain"].task)
+        team_id = temp_data[chat_id].get("team_id")
+        chain_order = temp_data[chat_id].get("chain_order")
+
+
+        team = get_team_by_id(team_id)
+        current_chain = get_current_chain(team_id, chain_order)
+
+        if not team or not current_chain:
+            bot.reply_to(message, QuestMessages.TEAM_NOT_FOUND)
+            state.delete()
+            return
+        
+        ok = check_task_code(message, team, current_chain.task)
+
+        if not ok:
+            return
+        
+        temp_data.pop(chat_id, None)
         state.delete()
 
     def check_task_code(message: Message, team: Team, current_team_task: Task):
@@ -135,18 +166,29 @@ def register_quest_commands(bot: TeleBot):
         next_order = team.current_chain_order + 1
         update_team(team_id=team.id, current_chain_order=next_order)
 
+
         all_members = get_user_ids_by_team(team.id)
         next_chain = get_current_chain(team.id, next_order)
+
+
         if not next_chain:
             for user_id in all_members:
                 try:
-                    bot.send_message(user_id, team.final_message or QuestMessages.QUEST_COMPLETED)
+                    bot.send_message(
+                        user_id, 
+                        team.final_message or QuestMessages.QUEST_COMPLETED,
+                        reply_markup=render_main_menu(check_admin(bot, message, silent=True), is_in_team=True)
+                    )
                 except Exception as e:
                     print(f"Не удалось отправить финальное сообщение пользователю {user_id}: {e}")    
             return
             
         task = next_chain.task
-        bot.reply_to(message, QuestMessages.TEAM_NEXT_TASK_MESSAGE)
+        bot.reply_to(
+            message,
+            QuestMessages.TEAM_NEXT_TASK_MESSAGE, 
+            reply_markup=render_main_menu(check_admin(bot, message, silent=True), is_in_team=True)
+        )
         send_task(who_solved, task)
 
         for user_id in all_members:
